pyproject.toml
[project]
name = "marketdatacollector"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "tastytrade>=10.3.0",
    "requests~=2.32.4"
]
